requirements.txt
MarkupSafe==1.0
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
click==6.7
colorize==1.1.0
dialogflow==0.6.0
dnspython==1.16.0
feedparser==5.2.1
Flask==0.12.3
fuzzywuzzy==0.17.0
gnewsclient==1.12
google-api-core==1.11.1
google-auth==1.6.3
googleapis-common-protos==1.6.0
grpcio==1.21.1
gunicorn==19.7.1
idna==2.8
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
protobuf==3.8.0
PTable==0.9.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyJWT==1.7.1
pymongo==3.8.0
PySocks==1.7.0
pytz==2017.2
requests==2.13.0
rsa==4.0
six==1.12.0
twilio==6.27.1
urllib3==1.25.3
weather-api==1.0.7
Werkzeug==0.12.1